// src/core/types.ts
/**
 * Enumeration of possible action types.
 * @enum {string}
 */
export enum ActionType {
    /**
     * Represents a publish action.
     */
    PUBLISH = 'publish',
    /**
     * Represents a request action.
     */
    REQUEST = 'request',
    /**
     * Represents a response action.
     */
    RESPONSE = 'response',
}